requirements.txt
Django==2.2.4
djangorestframework==3.10
sqreen
boto3
slackclient
requests